+++ main.js
@@ -59,7 +59,10 @@ function createWindow() {
 function createWorker() {
   // create hidden worker window
   const workerWindow = new BrowserWindow({
-    show: false
+    show: false,
+    webPreferences: {
+      nodeIntegration: true
+    }
   });
   // and load the autostart.html
   workerWindow.loadFile('autostart.html');
@@ -131,7 +134,25 @@ function checkTray() {
 // initialization and is ready to create browser windows.
 // Some APIs can only be used after this event occurs.
 app.whenReady().then(createWindow)
-app.whenReady().then(createWorker)
+app.whenReady().then(function () {
+  if (autostarted) {
+    createWorker();
+  } else {
+    win.webContents
+      .executeJavaScript('localStorage.getItem("settings");')
+      .then(result => {
+        if (result !== null) {
+          let settings = JSON.parse(result);
+          log.debug("Settings:");
+          log.debug(settings);
+          // show tray only if enabled
+          if (!settings[2].value) {
+            createWorker();
+          }
+        }
+      });
+  }
+})
 app.whenReady().then(checkTray)
 
 // Quit when all windows are closed, except on macOS. There, it's common